fix: fix reset sort and rang input
